+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Changes
 * Create pclass-instance data structure for HTN's
   - Fixes #146 Extend pclass-ctor-options to be strings OR references
   - Fixes #111 0-Arg main method not correctly used as the default root task
+* Further handle field-refs as part of the fix for #146
+* Add a command line option to set the dbg-println-level
 
 ### [0.6.2] - 2017-08-28